Implemented Auth.

+++ routes/greenhouse.php
@@ -10,6 +10,13 @@ $app->group('/greenhouse', function (RouteCollectorProxy $group) {
   //Function to list all the plants
   $group->get('/plants', function (Request $request, Response $response, $args) {
 
+    $decoded = $request->getAttribute("token");
+    $permissions = json_decode($decoded["scope"], true);
+
+    if(!$permissions['greenhouse_read']){
+      return $response->withStatus(401);
+    }
+
     $db = new Greenhouse();
     $data = json_encode($db->getPlants());
     $response->getBody()->write($data);
@@ -18,6 +25,13 @@ $app->group('/greenhouse', function (RouteCollectorProxy $group) {
 
   //Function to list all the plant species
   $group->get('/species', function (Request $request, Response $response, $args) {
+
+    $decoded = $request->getAttribute("token");
+    $permissions = json_decode($decoded["scope"], true);
+
+    if(!$permissions['greenhouse_read']){
+      return $response->withStatus(401);
+    }
 
     $db = new Greenhouse();
     $data = json_encode($db->getPlantSpecies());
@@ -28,6 +42,13 @@ $app->group('/greenhouse', function (RouteCollectorProxy $group) {
   //Function to list all the plant locations
   $group->get('/locations', function (Request $request, Response $response, $args) {
 
+    $decoded = $request->getAttribute("token");
+    $permissions = json_decode($decoded["scope"], true);
+
+    if(!$permissions['greenhouse_read']){
+      return $response->withStatus(401);
+    }
+
     $db = new Greenhouse();
     $data = json_encode($db->getPlantLocations());
     $response->getBody()->write($data);
@@ -37,9 +58,16 @@ $app->group('/greenhouse', function (RouteCollectorProxy $group) {
   //Function to add a new plant
   $group->post('/add', function (Request $request, Response $response, $args) {
 
+    $decoded = $request->getAttribute("token");
+    $permissions = json_decode($decoded["scope"], true);
+
+    if(!$permissions['greenhouse_write']){
+      return $response->withStatus(401);
+    }
+
     $db = new Greenhouse();
     $body = $request->getParsedBody();
-    
+
     $db->addPlant($body['plantName'], $body['plantSpecies'], $body['plantLocation']);
     
     return $response;
@@ -48,8 +76,18 @@ $app->group('/greenhouse', function (RouteCollectorProxy $group) {
   //Function to delete a plant
   $group->delete('/delete', function (Request $request, Response $response, $args) {
 
+    $decoded = $request->getAttribute("token");
+    $permissions = json_decode($decoded["scope"], true);
+
+    if(!$permissions['greenhouse_delete']){
+      return $response->withStatus(401);
+    }
+
     $db = new Greenhouse();
     $body = $request->getParsedBody();
+
+    $decoded = $request->getAttribute("token");
+    
     
     $db->deletePlant($body['plantID']);
